Added support for x64
git-svn-id: https://svn.xiph.org/trunk/oggdsf@15547 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ src/lib/codecs/flac/filters/dsfNativeFLACSource/NativeFLACSourceFilter.cpp
@@ -287,7 +287,7 @@ DWORD NativeFLACSourceFilter::ThreadProc(void)
 }
 
 
-::FLAC__StreamDecoderReadStatus NativeFLACSourceFilter::read_callback(FLAC__byte outBuffer[], unsigned int* outNumBytes) 
+::FLAC__StreamDecoderReadStatus NativeFLACSourceFilter::read_callback(FLAC__byte outBuffer[], size_t* outNumBytes) 
 {
 	const unsigned long BUFF_SIZE = 8192;
 	mInputFile.read((char*)outBuffer, BUFF_SIZE);
